refactor(api): split signup and createConsumer from user service
