+++ packages/api/tests/__helpers__/tokenHelper.ts
@@ -1,15 +1,15 @@
 import { RoleEnum } from "../../src/@enums/Roles";
-import userService from "../../src/modules/user/user.service";
 import userAuthService from "../../src/modules/user/services/auth/user.auth.service";
 import userRolesService from "../../src/modules/user/services/roles/user.roles.service";
 import userActivationService from "../../src/modules/user/services/activation/user.activation.service";
 import userCrudService from "../../src/modules/user/services/crud/user.crud.service";
+import userConsumerService from "../../src/modules/user/services/consumer/user.consumer.service";
 
 const getToken = async (role = RoleEnum.user) => {
     const email = `${role}@beta.gouv.fr`;
     let user = await userCrudService.findByEmail(email);
     if (!user) {
-        if (role == RoleEnum.consumer) user = await userService.createConsumer({ email: email });
+        if (role == RoleEnum.consumer) user = await userConsumerService.createConsumer({ email: email });
         else user = await userCrudService.createUser({ email });
     }
